server/requirements.txt
django==2.2
djangorestframework==3.9
djangorestframework-jsonp
psycopg2
joblib
numpy
pandas
uwsgi
scikit-learn==0.21.2
